+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - HTTP COM only supports powershell agent and uses an older COM object that isn't used often
   - OneDrive has new APIs and Microsoft hs made registration harder. May return in the future with revisions.
   - Dropbox has new APIs and may return in the future with revisions.
+-   Revised the staging process for agents. Session IDs are provided by the server and all packets are wrapped in routing packets.
 
 ## [6.0.0-beta3] - 2025-03-02